Add files via upload
This dataset is a collection of responses to an anonymous questionnaire answered by 10 people. The questionnaire includes questions on various aspects related to mental health, modern life, and factors that harm the mental health of mothers, fathers, and children specifically. The dataset contains 13 columns, including demographic information such as age and gender, years of formal education, suggested leading and second-leading causes of mental health problems, mechanisms through which those factors lead to mental health problems, favorite and least favorite features of modern life, and factors that harm the mental health of mothers, fathers, and children specifically. This dataset can be used for various research or analysis related to mental health and its various factors.