Add files via upload
In this project, I surveyed 10 participants about their degree of agreement with three different approaches to feminism, namely difference-denying feminism, difference-removing feminism, and difference-honoring feminism. I asked them to rate their agreement with each approach on a scale of 1 to 10, with 1 being "strongly disagree" and 10 being "strongly agree". Additionally, I asked them which of the three approaches they believe has had the most cultural and political impact in the past few decades and whether they are or have been a mother. The survey data was compiled into a CSV file with columns for the answers to each question.